requirements.txt
boto3>=1.28.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.23